chore: Make HTML functions more modular
- Split code into several files
- Add translations.ts
