+++ src/translations.ts
@@ -1,0 +1,11 @@
+// translations.ts
+export const translations = {
+  noMatch: 'Ingen träff hos Vivino!',
+  notOnBottle: 'Vinet är inte på flaska',
+  loading: 'Laddar...',
+  linkToVivino: 'Länk till Vivino',
+  rating: 'Betyg',
+  votes: 'röster',
+  of: 'av'
+  // Add more translations as needed
+}
